web_panel/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.3
Flask-WTF==1.1.1
WTForms==3.0.1
Werkzeug==2.3.7
Flask-Limiter==3.5.0
Flask-JWT-Extended==4.5.3
gunicorn==21.2.0
python-dotenv==1.0.0
telethon==1.42.0
cryptg==0.4.0